refactor: change code structure
